refactor(main): extract attach/detach into host/{attach,detach}
