+++ src/main/host/detach.ts
@@ -1,0 +1,328 @@
+import { dialog, ipcMain } from 'electron'
+import type { BrowserWindow, WebContentsView } from 'electron'
+import * as ipc from '../lib/ipc'
+import { COMFY_BG } from '../lib/theme'
+import { _unregisterExtraBroadcastTarget } from '../lib/ipc/shared'
+import { comfyWindows } from './registry'
+import type { ComfyWindowEntry } from './registry'
+import {
+  applyChooserHostTheme,
+  CHOOSER_HOST_TITLE_TEXT,
+  CHOOSER_HOST_WINDOW_TITLE,
+  openChooserHostWindow,
+} from './createHostWindow'
+
+/** Late-bound dependency on the panelView constructor that still
+ *  lives in `index.ts` (until P0.9). */
+export interface DetachFactories {
+  ensurePanelView: (
+    windowKey: number,
+    entry: ComfyWindowEntry,
+    initialPanel: 'chooser',
+  ) => WebContentsView
+}
+
+let factories: DetachFactories | null = null
+
+export function setDetachFactories(opts: DetachFactories): void {
+  factories = opts
+}
+
+function getFactories(): DetachFactories {
+  if (!factories) {
+    throw new Error('setDetachFactories must be called before detach')
+  }
+  return factories
+}
+
+/**
+ * WeakSet of host windows whose `close` should skip the panel-renderer
+ * consult and tear down immediately. The bulk-close confirm dialog
+ * already lists in-progress operations / sessions / downloads, so the
+ * per-window prompt is redundant noise once the user has confirmed
+ * the bulk close. Used by `confirmAndCloseAllHostWindows` (here) and
+ * `returnToDashboard` (also here); the close handler in
+ * `createHostWindow` reads it via `setHostWindowFactories`.
+ */
+export const preClearedClose = new WeakSet<BrowserWindow>()
+
+/**
+ * Main consults the panel renderer before tearing down a host
+ * window so a Tier 2 progress / Tier 3 takeover overlay can
+ * prompt the user to confirm cancellation via the standardised
+ * cancel-prompt copy. Returns true when the renderer cleared the
+ * close (no overlay open, or the user confirmed cancellation),
+ * false when the renderer aborted (user dismissed the prompt).
+ *
+ * Falls back to "cleared" when the panelView is missing (no panel
+ * has been mounted yet — nothing to lose), the webContents is
+ * destroyed (already torn down), the renderer doesn't ack receipt
+ * of the request within 2s (hung renderer), or the underlying
+ * webContents goes away (render-process-gone / destroyed).
+ *
+ * Important: once the renderer acks receipt we wait INDEFINITELY for
+ * the actual response. The renderer might be showing a confirmation
+ * modal that the user takes their time on; an extra fixed timeout
+ * here would force-close the window out from under that prompt
+ * (which was the bug observed when a sub-5s prompt-response window
+ * triggered an unconfirmed close).
+ */
+export async function consultPanelRendererClose(
+  panelView: WebContentsView | null | undefined,
+): Promise<boolean> {
+  if (!panelView || panelView.webContents.isDestroyed()) return true
+  return new Promise<boolean>((resolve) => {
+    const requestId = `close-${Date.now()}-${Math.random().toString(36).slice(2)}`
+    let settled = false
+    let acked = false
+    const cleanup = (): void => {
+      ipcMain.off('comfy-window:request-close-ack', onAck)
+      ipcMain.off('comfy-window:request-close-response', onResponse)
+      if (!panelView.webContents.isDestroyed()) {
+        panelView.webContents.off('render-process-gone', onCrash)
+        panelView.webContents.off('destroyed', onCrash)
+      }
+    }
+    const onAck = (
+      event: Electron.IpcMainEvent,
+      payload: { requestId?: string } | undefined,
+    ): void => {
+      if (event.sender !== panelView.webContents) return
+      if (payload?.requestId !== requestId) return
+      acked = true
+    }
+    const onResponse = (
+      event: Electron.IpcMainEvent,
+      payload: { requestId?: string; cleared?: boolean } | undefined,
+    ): void => {
+      if (event.sender !== panelView.webContents) return
+      if (payload?.requestId !== requestId) return
+      if (settled) return
+      settled = true
+      cleanup()
+      resolve(!!payload?.cleared)
+    }
+    const onCrash = (): void => {
+      if (settled) return
+      settled = true
+      cleanup()
+      resolve(true)
+    }
+    ipcMain.on('comfy-window:request-close-ack', onAck)
+    ipcMain.on('comfy-window:request-close-response', onResponse)
+    panelView.webContents.on('render-process-gone', onCrash)
+    panelView.webContents.on('destroyed', onCrash)
+    try {
+      panelView.webContents.send('comfy-window:request-close', { requestId })
+    } catch {
+      settled = true
+      cleanup()
+      resolve(true)
+      return
+    }
+    // Hung-renderer safety: only fires if we never got the ack. Once
+    // the renderer acks receipt we trust it to either reply or have
+    // its webContents torn down (render-process-gone covers that).
+    setTimeout(() => {
+      if (settled || acked) return
+      settled = true
+      cleanup()
+      resolve(true)
+    }, 2000)
+  })
+}
+
+/**
+ * Close every host window (install-backed and chooser hosts alike) but
+ * leave the app / tray alive. Bound to the File menu's "Close All
+ * Windows" entry. Each window's existing `close` handler runs the
+ * full teardown (`stopRunning` + webContents close + window.destroy),
+ * so we just dispatch `close()` and let those handlers do the work
+ * — the handlers also consult the panel renderer unless the window
+ * is already in `preClearedClose`. Snapshot the entry list
+ * first so the iteration isn't affected by `closed` callbacks that
+ * delete from the `comfyWindows` map mid-loop.
+ */
+export function closeAllHostWindows(): void {
+  const entries = Array.from(comfyWindows.values())
+  for (const entry of entries) {
+    if (!entry.window.isDestroyed()) entry.window.close()
+  }
+}
+
+/**
+ * File menu's "Return to Dashboard" entry. Closes the install-backed
+ * host window and opens a chooser host window at the same bounds.
+ *
+ * In-place flip via `entry.detachInstall()` is currently disabled
+ * — too many edge-case bugs around the in-place swap. The close+open
+ * swap pays a visible flicker but exercises the same close-handler
+ * teardown that production has used since main, which is the
+ * codepath we trust right now. See
+ * docs/window-mode-unification-revert.md.
+ */
+export async function returnToDashboard(parentEntryId: number): Promise<void> {
+  const entry = comfyWindows.get(parentEntryId)
+  if (!entry || entry.installationId === null || entry.window.isDestroyed()) return
+  const cleared = await consultPanelRendererClose(entry.panelView)
+  if (!cleared) return
+  if (entry.window.isDestroyed()) return
+  preClearedClose.add(entry.window)
+  const bounds = entry.window.getBounds()
+  const wasMaximized = entry.window.isMaximized()
+  const chooserWindow = openChooserHostWindow()
+  if (!chooserWindow.isDestroyed()) {
+    if (wasMaximized) {
+      chooserWindow.maximize()
+    } else {
+      chooserWindow.setBounds(bounds)
+    }
+  }
+  entry.window.close()
+}
+
+/**
+ * Confirm a `closeAllHostWindows()` dispatch when more than one host
+ * window is open. The dialog lists the open windows by title (so the
+ * user can see what's about to close) and any active operations that
+ * will be cancelled — running ComfyUI sessions, in-progress
+ * installs / updates, active model downloads — pulled from the same
+ * `getActiveDetails()` helper. With one or zero windows the close
+ * happens straight through with no prompt.
+ */
+export async function confirmAndCloseAllHostWindows(
+  parentWindow: BrowserWindow | null,
+): Promise<void> {
+  const entries = Array.from(comfyWindows.values()).filter((e) => !e.window.isDestroyed())
+  if (entries.length <= 1) {
+    closeAllHostWindows()
+    return
+  }
+  const titles = entries.map((e) => e.window.getTitle() || 'Untitled window')
+  const detailLines: string[] = ['Open windows:', ...titles.map((t) => `  • ${t}`)]
+  if (ipc.hasActiveOperations()) {
+    try {
+      const items = await ipc.getActiveDetails()
+      const sessions = items.filter((i) => i.type === 'session').map((i) => i.name)
+      const operations = items.filter((i) => i.type === 'operation').map((i) => i.name)
+      const downloads = items.filter((i) => i.type === 'download').map((i) => i.name)
+      if (sessions.length > 0) {
+        detailLines.push('', 'Running ComfyUI:', ...sessions.map((n) => `  • ${n}`))
+      }
+      if (operations.length > 0) {
+        detailLines.push('', 'In-progress operations:', ...operations.map((n) => `  • ${n}`))
+      }
+      if (downloads.length > 0) {
+        detailLines.push('', 'Active downloads:', ...downloads.map((n) => `  • ${n}`))
+      }
+    } catch {
+      // If active-detail collection ever throws, fall back to just the
+      // window list — the user still sees what's about to close.
+    }
+  }
+  const opts: Electron.MessageBoxOptions = {
+    type: 'question',
+    buttons: ['Close All', 'Cancel'],
+    defaultId: 1,
+    cancelId: 1,
+    title: 'Close All Windows',
+    message: `Close ${entries.length} open windows?`,
+    detail: detailLines.join('\n'),
+  }
+  const result = parentWindow && !parentWindow.isDestroyed()
+    ? await dialog.showMessageBox(parentWindow, opts)
+    : await dialog.showMessageBox(opts)
+  if (result.response === 0) {
+    // The global dialog already lists in-progress ops / sessions /
+    // downloads, so the per-window tier-aware prompt would be
+    // redundant after the user confirmed the bulk close. Pre-clear
+    // every entry so each window's `close` handler skips its own
+    // consult and tears down immediately.
+    for (const entry of entries) preClearedClose.add(entry.window)
+    closeAllHostWindows()
+  }
+}
+
+/**
+ * Flip an install-backed host window in place to install-less
+ * (chooser) mode. The symmetric undo to `attachInstall()`. Bound
+ * onto `entry.detachInstall` by `createHostWindow()`; the
+ * underscore-prefixed name signals that callers should invoke
+ * `entry.detachInstall()` rather than this freestanding helper
+ * directly.
+ *
+ * Steps:
+ *   1. Runs `entry._installCleanup()` — `attachInstall()`'s stashed
+ *      undo: off all install-bound comfyContents listeners, cancel
+ *      the fail-retry timer, ipc.stopRunning the running session,
+ *      clear the install-keyed maps + the secondary index, and reset
+ *      `entry.installationId` / `entry.comfyUrl`.
+ *   2. Navigates the comfyView to `about:blank` so the loaded
+ *      ComfyUI page is unloaded (releases its renderer process). The
+ *      comfyView is kept alive (not destroyed) so the host can be
+ *      re-attached later without rebuilding.
+ *   3. Resets the title-bar identity (`titleBarText` →
+ *      `'Desktop 2.0 Beta'`, `sourceCategory` → `null`) and pushes
+ *      to the live title-bar.
+ *   4. Resets the OS-level window title.
+ *   5. Re-paints the title bar to the launcher-theme surface
+ *      (chooser hosts derive their theme from the launcher setting,
+ *      not from a ComfyUI frontend).
+ *   6. Resets `entry.activePanel` to `'comfy'` (which now resolves
+ *      to the chooser body via `computeBodyMode`) and ensures a
+ *      panelView with the chooser body exists.
+ *   7. Calls `entry.layoutViews()` so the chooser body becomes
+ *      visible immediately.
+ *
+ * No-op when the entry is already install-less (no install backing
+ * to detach). Does not destroy the comfyView or the BrowserWindow
+ * — see the close handler in `createHostWindow()` for the destroy
+ * path.
+ */
+export function _detachInstallImpl(entry: ComfyWindowEntry): void {
+  if (entry.installationId === null) return
+  if (entry.window.isDestroyed()) return
+  const fx = getFactories()
+
+  // Symmetric undo of attachInstall (listeners, maps, stopRunning, etc).
+  entry._installCleanup?.()
+
+  // Release the ComfyUI page; the view is kept alive for re-attach.
+  if (!entry.comfyView.webContents.isDestroyed()) {
+    void entry.comfyView.webContents.loadURL('about:blank').catch(() => {})
+    entry.comfyView.setBackgroundColor(COMFY_BG)
+  }
+
+  // Flip title-bar identity back to chooser-host shape.
+  entry.titleBarText = CHOOSER_HOST_TITLE_TEXT
+  entry.sourceCategory = null
+  if (!entry.titleBarView.webContents.isDestroyed()) {
+    entry.titleBarView.webContents.send('comfy-titlebar:title-changed', entry.titleBarText)
+    entry.titleBarView.webContents.send('comfy-titlebar:source-category-changed', null)
+  }
+  entry.window.setTitle(CHOOSER_HOST_WINDOW_TITLE)
+  applyChooserHostTheme(entry)
+
+  // Reset nav state to the comfy pill (chooser body for install-less hosts).
+  entry.activePanel = 'comfy'
+  if (!entry.titleBarView.webContents.isDestroyed()) {
+    entry.titleBarView.webContents.send('comfy-titlebar:panel-changed', 'comfy')
+  }
+
+  // Tear down the install-backed PanelApp and remount fresh in chooser mode.
+  // Preserves no per-install state (overlays, activePanel, installationId
+  // URL param) across the detach.
+  if (entry.panelView) {
+    const oldPanel = entry.panelView
+    entry.panelView = null
+    if (!oldPanel.webContents.isDestroyed()) {
+      _unregisterExtraBroadcastTarget(oldPanel.webContents)
+      oldPanel.webContents.close()
+    }
+    if (!entry.window.isDestroyed()) {
+      try { entry.window.contentView.removeChildView(oldPanel) } catch {}
+    }
+  }
+  fx.ensurePanelView(entry.windowKey, entry, 'chooser')
+  entry.layoutViews()
+}
